reorganize kernelspec
