+++ jupyter_server/serverapp.py
@@ -284,7 +284,7 @@ class ServerWebApplication(web.Application):
         handlers.extend(load_handlers('jupyter_server.view.handlers'))
         handlers.extend(load_handlers('jupyter_server.nbconvert.handlers'))
         handlers.extend(load_handlers('jupyter_server.bundler.handlers'))
-        handlers.extend(load_handlers('jupyter_server.kernelspecs.handlers'))
+        handlers.extend(load_handlers('jupyter_server.services.kernelspecs.handlers'))
         handlers.extend(load_handlers('jupyter_server.edit.handlers'))
         handlers.extend(load_handlers('jupyter_server.services.api.handlers'))
         handlers.extend(load_handlers('jupyter_server.services.config.handlers'))
@@ -292,7 +292,7 @@ class ServerWebApplication(web.Application):
         handlers.extend(load_handlers('jupyter_server.services.contents.handlers'))
         handlers.extend(load_handlers('jupyter_server.services.sessions.handlers'))
         handlers.extend(load_handlers('jupyter_server.services.nbconvert.handlers'))
-        handlers.extend(load_handlers('jupyter_server.services.kernelspecs.handlers'))
+        handlers.extend(load_handlers('jupyter_server.services.kernelspecs.api_handlers'))
         handlers.extend(load_handlers('jupyter_server.services.security.handlers'))
         handlers.extend(load_handlers('jupyter_server.services.shutdown'))
         handlers.extend(settings['contents_manager'].get_extra_handlers())
